# Vectorize `basic_string::find`