Resolves #5036

**⚠️ Need to decide whether this CRT replacing vectorization is needed ⚠️** 

# 🏁 Race against the runtime!

`basic_string::find` is already vectorized for 8 and 16 bit characters via calling `memchr` and `wmemchr`

`memchr` shows good results for smaller positions, but if the distance from the beginning to the result is long the `memchr` results are noticeably worse. There's strong indication that such a results are due to not using AVX in `memchr`.

`wmemchr` is not fast at all. It was promised to be fast in a new CRT in #4873 though.

For larger characters the vectorization is novel, there's no CRT implementation, but these characters are rare anyway.

# ⏱️ Benchmark results

Benchmark                                                              |      main    |       this    
-----------------------------------------------------------------------|--------------|---------------
bm<char, not_highly_aligned_allocator, Op::StringFind>/8021/3056       |   89.6 ns    |    49.3 ns    
bm<char, not_highly_aligned_allocator, Op::StringFind>/63/62           |   11.3 ns    |    3.70 ns    
bm<char, not_highly_aligned_allocator, Op::StringFind>/31/30           |   9.90 ns    |    8.70 ns    
bm<char, not_highly_aligned_allocator, Op::StringFind>/15/14           |   4.94 ns    |    7.10 ns    
bm<char, not_highly_aligned_allocator, Op::StringFind>/7/6             |   2.83 ns    |    3.49 ns    
bm<char, highly_aligned_allocator, Op::StringFind>/8021/3056           |   78.0 ns    |    48.0 ns    
bm<char, highly_aligned_allocator, Op::StringFind>/63/62               |   12.4 ns    |    4.11 ns    
bm<char, highly_aligned_allocator, Op::StringFind>/31/30               |   10.2 ns    |    7.85 ns    
bm<char, highly_aligned_allocator, Op::StringFind>/15/14               |   5.18 ns    |    7.42 ns    
bm<char, highly_aligned_allocator, Op::StringFind>/7/6                 |   2.76 ns    |    3.42 ns    
bm<wchar_t, not_highly_aligned_allocator, Op::StringFind>/8021/3056    |    754 ns    |    85.3 ns    
bm<wchar_t, not_highly_aligned_allocator, Op::StringFind>/63/62        |   31.1 ns    |    3.84 ns    
bm<wchar_t, not_highly_aligned_allocator, Op::StringFind>/31/30        |   16.2 ns    |    2.92 ns    
bm<wchar_t, not_highly_aligned_allocator, Op::StringFind>/15/14        |   4.65 ns    |    3.66 ns    
bm<wchar_t, not_highly_aligned_allocator, Op::StringFind>/7/6          |   2.45 ns    |    2.93 ns    
bm<char32_t, not_highly_aligned_allocator, Op::StringFind>/8021/3056   |    749 ns    |     159 ns    
bm<char32_t, not_highly_aligned_allocator, Op::StringFind>/63/62       |   18.5 ns    |    5.25 ns    
bm<char32_t, not_highly_aligned_allocator, Op::StringFind>/31/30       |   9.60 ns    |    3.55 ns    
bm<char32_t, not_highly_aligned_allocator, Op::StringFind>/15/14       |   6.42 ns    |    2.91 ns    
bm<char32_t, not_highly_aligned_allocator, Op::StringFind>/7/6         |   2.80 ns    |    2.95 ns