@StephanTLavavej 2024-12-05
[Detective Beans and the Case of the Missing `char`!](https://www.simonandschuster.com/books/Detective-Beans/Li-Chen/9781524882587) :cat: :detective: :mag_right: